refactor(agent-core): split the files PEER-007 pushed past their size baselines
The attribution change added a field or two to six files that were each sitting exactly
at a frozen size baseline, so `file-size` refused the push. The rule's answer to that is
"split instead of extending", and taking it seriously turned a line-counting problem into
five modules that should have existed anyway:

- `packages/agent-core/src/interfaces/run-options.ts` — `IRunOptions` is the per-RUN contract, distinct from the
  agent's standing `IAgentConfig`. Moving it (with the two execution-event types, whose
  only consumer is `onExecutionEvent`) takes `packages/agent-core/src/interfaces/agent.ts` from 308 lines to 242
  and off the baseline list entirely.
- `packages/agent-core/src/services/execution-user-message.ts` — what metadata a stored user message carries.
- `packages/agent-transport-tui/src/RoleLabel.tsx` — PEER-007 gave the `user` case a second thing to decide,
  so the label stopped being a constant per role and became its own responsibility.
- `packages/agent-transport-tui/src/attributed-user-echo.ts` — building the attributed echo, guard included.
- `packages/agent-session/src/session-run-options.ts` + `ISessionRunOptions` in `packages/agent-session/src/session-types.ts` —
  the two per-turn options were spread one at a time at each call site; now they are a
  group, which is what they are.

`promptTurnAttribution` in `packages/agent-framework/src/interactive/interactive-session-prompt.ts` does the same joining for the
prompt context, and owns the rule that the OWNER driver is elided rather than stamped on
every message.

The baseline is regenerated in this change because three files ended up BELOW their frozen
counts and the scan asks for the gain to be locked in the same commit. Every row in that
diff moves one direction — `packages/agent-core/src/services/execution-service.ts` 305→304, `packages/agent-session/src/session-run.ts` 323→322, and
`packages/agent-core/src/interfaces/agent.ts` dropped out at 242. Nothing was adopted upward, which is the check that
distinguishes locking a refactor from quietly adopting debt.

Also records two progress-report-quantification findings of my own, both a bare "3/4"
without the percentage — the second inside the very message acknowledging the first.
Transcripts are append-only, so the scan's contract is to record them, not to edit history.

Verified: file-size and progress-report-quantification both green; 3,302 tests pass across
agent-core, agent-session, agent-framework and agent-transport-tui; workspace typecheck clean.
