+++ scripts/harness/file-size-baseline.json
@@ -11,12 +11,11 @@
   "packages/agent-core/src/abstracts/abstract-workflow-converter.ts": 319,
   "packages/agent-core/src/core/robota.ts": 411,
   "packages/agent-core/src/index.ts": 313,
-  "packages/agent-core/src/interfaces/agent.ts": 308,
   "packages/agent-core/src/managers/agent-factory.ts": 301,
   "packages/agent-core/src/plugins/event-emitter-plugin.ts": 310,
   "packages/agent-core/src/services/execution-round.ts": 304,
   "packages/agent-core/src/services/execution-service-helpers.ts": 313,
-  "packages/agent-core/src/services/execution-service.ts": 305,
+  "packages/agent-core/src/services/execution-service.ts": 304,
   "packages/agent-executor/src/background-tasks/background-task-manager.ts": 398,
   "packages/agent-framework/src/checkpoints/edit-checkpoint-store.ts": 390,
   "packages/agent-framework/src/interactive/interactive-session-background-tracker.ts": 334,
@@ -40,7 +39,7 @@
   "packages/agent-plugin/src/usage/usage-plugin.ts": 301,
   "packages/agent-plugin/src/webhook/webhook-plugin.ts": 302,
   "packages/agent-provider-openai-compatible/src/qwen/responses-parser.ts": 302,
-  "packages/agent-session/src/session-run.ts": 323,
+  "packages/agent-session/src/session-run.ts": 322,
   "packages/agent-session/src/session.ts": 316,
   "packages/agent-transport-tui/src/App.tsx": 605,
   "packages/agent-transport-tui/src/InputArea.tsx": 318,
